refactor(supernova): introduced the fetch to the useQuery classes
