--- apps/supernova/src/api/alerts.js
@@ -1,53 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Greenhouse contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { sortAlerts, countAlerts } from "../lib/utils"
-let compareAlertString
-
-export const fetchAlerts = async (endpoint) => {
-  try {
-    const response = await fetch(`${endpoint}/alerts`)
-
-    if (!response.ok) {
-      // Parse the error object from the response body
-      const errorObject = await response.json().catch(() => {
-        throw new Error(`Unexpected error: Unable to parse error response.`)
-      })
-
-      // Throw the error object directly
-      throw errorObject
-    }
-
-    const items = await response.json() // Parse JSON data
-
-    let alerts = sortAlerts(items)
-
-    // copy additional filter options to labels for easier filter selection
-    // because the alert object is nested this makes it a lot easier to filter, since we only use what is present in alert.labels
-    alerts.forEach((alert) => {
-      if (alert.labels) {
-        alert.labels.status = alert.status?.state
-      }
-    })
-
-    // check if new loaded alerts are different from the last response
-    const newCompareString = JSON.stringify(alerts)
-    if (compareAlertString !== newCompareString) {
-      compareAlertString = newCompareString
-
-      // inform listener to receive new alerts
-      return {
-        alerts: alerts,
-        counts: countAlerts(alerts),
-      }
-    } else {
-      // return null when nothing new
-      return null
-    }
-  } catch (error) {
-    console.error(error)
-    throw error // Let React Query handle the error
-  }
-}
