+++ apps/supernova/src/hooks/useSilencesQuery.js
@@ -6,8 +6,6 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query"
 import { useGlobalsApiEndpoint } from "../components/StoreProvider"
 
-import { fetchSilences } from "../api/silences"
-
 export const useSilencesQuery = (key, { options = {} } = {}) => {
   const queryClient = useQueryClient()
   const endpoint = useGlobalsApiEndpoint()
@@ -28,4 +26,30 @@ export const useSilencesQuery = (key, { options = {} } = {}) => {
     },
     ...options, // Include any additional options like staleTime, retry, etc.
   })
+}
+
+const fetchSilences = async (endpoint) => {
+  try {
+    const response = await fetch(`${endpoint}/silences`)
+
+    if (!response.ok) {
+      // Parse the error object from the response body
+      const errorObject = await response.json().catch(() => {
+        throw new Error(`Unexpected error: Unable to parse error response.`)
+      })
+
+      // Throw the error object directly
+      throw errorObject
+    }
+
+    const items = await response.json() // Parse JSON data
+
+    // Return the structured result
+    return {
+      silences: items,
+    }
+  } catch (error) {
+    console.error(error)
+    throw error // Let React Query handle the error
+  }
 }
